More Progress
The kv file was loading twice for some reason so i had to rename it.

Basically meshed the blueprint that i had with the Kivy stuff. Added the middle layout where the cards get placed. I haven't begun doing any game stuff or anything but right now, two players are made, and they both have cards. They can toggle their cards and whichever ones are toggled when they select the field get removed from player and added to field. This process is not 100% complete. Prob need another hour for this.